--- a/src/_kepler.pyx
+++ b/src/_kepler.pyx
@@ -3,7 +3,8 @@ from __future__ import absolute_import
 cimport numpy as np
 import numpy as np
 import cython
-
+from libc.math cimport sin
+from libc.math cimport cos
 
 # if you want to use the Numpy-C-API from Cython
 # (not strictly necessary for this example, but good practice)
@@ -13,7 +14,7 @@ np.import_array()
 # arguments and returns a double
 cdef extern from "kepler.c":
     double kepler(double M, double e)
-    double rv_drive(double t, double per, double tp, double e, double om, double k )
+    double rv_drive(double t, double per, double tp, double e, double cosom, double sinom, double k )
 
 DTYPE = np.float64
 ctypedef np.float64_t DTYPE_t
@@ -40,8 +41,10 @@ def rv_drive_array(np.ndarray[DTYPE_t, ndim=1] t, double per, double tp,
     size = t.shape[0]
 
     cdef np.ndarray[DTYPE_t, ndim=1] rv = t.copy()
+    cdef double cosom = cos(om)
+    cdef double sinom = sin(om)
     for i in range(size):
-        rv[i] = rv_drive(t[i], per, tp, e, om, k)
+        rv[i] = rv_drive(t[i], per, tp, e, cosom, sinom, k)
 
     return rv