I implemented Tim Brandt's suggestion for cutting out some of the trig evaluations in RV drive.

Before
```

ECCENTRICITY = 0.1

Profiling pure C code for an RV time series with 10 observations
Ran 5000 model calculations in 0.021 seconds
Profiling Python code for an RV time series with 10 observations
Ran 5000 model calculations in 0.297 seconds
The C version runs 14.27 times faster

Profiling pure C code for an RV time series with 30 observations
Ran 5000 model calculations in 0.044 seconds
Profiling Python code for an RV time series with 30 observations
Ran 5000 model calculations in 0.319 seconds
The C version runs  7.28 times faster

Profiling pure C code for an RV time series with 100 observations
Ran 5000 model calculations in 0.139 seconds
Profiling Python code for an RV time series with 100 observations
Ran 5000 model calculations in 0.385 seconds
The C version runs  2.77 times faster

Profiling pure C code for an RV time series with 300 observations
Ran 5000 model calculations in 0.419 seconds
Profiling Python code for an RV time series with 300 observations
Ran 5000 model calculations in 0.515 seconds
The C version runs  1.23 times faster

Profiling pure C code for an RV time series with 1000 observations
Ran 5000 model calculations in 1.398 seconds
Profiling Python code for an RV time series with 1000 observations
Ran 5000 model calculations in 0.967 seconds
The C version runs  0.69 times faster

ECCENTRICITY = 0.7

Profiling pure C code for an RV time series with 30 observations
Ran 5000 model calculations in 0.052 seconds
Profiling Python code for an RV time series with 30 observations
Ran 5000 model calculations in 0.332 seconds
The C version runs  6.45 times faster

```

brandt-upgrade

```

ECCENTRICITY = 0.1

Profiling pure C code for an RV time series with 10 observations
Ran 5000 model calculations in 0.022 seconds
Profiling Python code for an RV time series with 10 observations
Ran 5000 model calculations in 0.333 seconds
The C version runs 15.27 times faster

Profiling pure C code for an RV time series with 30 observations
Ran 5000 model calculations in 0.038 seconds
Profiling Python code for an RV time series with 30 observations
Ran 5000 model calculations in 0.344 seconds
The C version runs  8.99 times faster

Profiling pure C code for an RV time series with 100 observations
Ran 5000 model calculations in 0.116 seconds
Profiling Python code for an RV time series with 100 observations
Ran 5000 model calculations in 0.378 seconds
The C version runs  3.27 times faster

Profiling pure C code for an RV time series with 300 observations
Ran 5000 model calculations in 0.318 seconds
Profiling Python code for an RV time series with 300 observations
Ran 5000 model calculations in 0.532 seconds
The C version runs  1.67 times faster

Profiling pure C code for an RV time series with 1000 observations
Ran 5000 model calculations in 1.031 seconds
Profiling Python code for an RV time series with 1000 observations
Ran 5000 model calculations in 0.994 seconds
The C version runs  0.96 times faster

ECCENTRICITY = 0.7

Profiling pure C code for an RV time series with 30 observations
Ran 5000 model calculations in 0.043 seconds
Profiling Python code for an RV time series with 30 observations
Ran 5000 model calculations in 0.326 seconds
The C version runs  7.55 times faster
```

I didn't see much of an improvement. I wonder if this is due to some massive overhead shipping the data between c and python. Also strange how the python/numpy version works faster with longer arrays. Perhaps the vast majority of the time is spent solving Kepler's equation, so these gains are not really noticeable.