@bjfultn 2019-06-11
Thanks for implementing this so fast @petigura! It looks like it works, so do you see any reason not to push it even if its only a marginal performance gain?