@petigura 2019-06-11
I want to understand why the performance boost is so marginal. We removed three trig evaluations per rv_drive call. Has a minimum of 5 trig calls if it goes through one iteration and 8 if it goes through two. It usually converges after two. Thus the ~75% runtime makes sense given that the number of trig calls has gone from 9 or 12 to 6 or 9.