@bjfultn 2019-11-14
I'm going to go ahead and merge this PR into `next-release`. It might not produce the performance enhancements that we were expecting, but there is some improvement and it certainly doesn't hurt.