@bjfultn 2019-11-15
Yes, the RV models are identical to floating point precision.

![ecc_diff](https://user-images.githubusercontent.com/1724542/68960605-579a5b80-0785-11ea-9a69-14879820785d.png)
![circ_diff](https://user-images.githubusercontent.com/1724542/68960610-59fcb580-0785-11ea-9b0a-6ca82485b39d.png)